Merge in RAD/fmdtools from hotfix/RAD-542-issues-for-fmdtools-2.2.1 to dev

* commit '8c8cc3b0502044058c141be6952b58dd24045a00':
  Adapting Component Architecture to new capabilities
  Update drone_mdl_hierarchical.py
  Migrating Component Architecture Examples
  Numpy types in states now show str() instead of repr()
  Containers now adapt type on assignment
  Objects in States/Containers can now be tracked
  Some syntax fixes

+++ examples/eps/eps.py
@@ -35,11 +35,12 @@ from fmdtools.define.container.mode import Mode
 from fmdtools.define.architecture.function import FunctionArchitecture
 from fmdtools.define.container.state import State
 from fmdtools.define.flow.base import Flow
+import numpy as np
 
 
 class GenericState(State):
-    rate: float = 1.0
-    effort: float = 1.0
+    rate: np.float64 = 1.0
+    effort: np.float64 = 1.0
 
 
 class GenericFlow(Flow):
@@ -48,7 +49,7 @@ class GenericFlow(Flow):
 
 
 class SigState(State):
-    value: float = 1.0
+    value: np.float64 = 1.0
 
 
 class Signal(Flow):
